+++ bpack/np.py
@@ -1,5 +1,6 @@
 """Numpy based codec for binary data structures."""
 
+import math
 import collections
 
 import numpy as np
@@ -16,7 +17,7 @@ from .descriptors import (
 
 __all__ = [
     'Decoder', 'decoder', 'BACKEND_NAME', 'BACKEND_TYPE',
-    'descriptor_to_dtype',
+    'descriptor_to_dtype', 'unpackbits',
 ]
 
 
@@ -153,3 +154,77 @@ class Decoder(bpack.codecs.Decoder):
 
 
 decoder = bpack.codecs.make_codec_decorator(Decoder)
+
+
+def unpackbits(data: bytes, bits_per_sample: int, signed: bool = False):
+    """Unpack packed (integer) values form a string of bytes.
+
+    Takes in input a string of bytes in which (integer) samples have been
+    stored using ``bits_per_sample`` bit for each sample, and returns
+    the sequence of corresponding Python integers.
+
+    Example::
+
+                 3 bytes                            4 samples
+
+      |------|------|------|------| --> [samp_1, samp_2, samp_3, samp_4]
+
+      4 samples (6 bits per sample)
+
+    If ``signed`` is set to True integers are assumed to be stored as
+    signed integers.
+    """
+    # TODO: support byteorder: str = ''
+    # if byteprder:
+    #     raise NotImplementedError(f'byteorder = "{byteprder}"')
+
+    if bits_per_sample > 8:
+        # TODO: support bits_per_sample > 8
+        raise NotImplementedError(f'bits_per_sample = {bits_per_sample}')
+    elif bits_per_sample == 8:
+        dtype = np.dtype('int8') if signed else np.dtype('uint8')
+        return np.frombuffer(data, dtype=dtype)
+    nbits = len(data) * 8
+    assert nbits % bits_per_sample == 0
+    samples = nbits // bits_per_sample
+
+    chunksize = bits_per_sample // math.gcd(bits_per_sample, 8)     # bytes
+    samples_per_chunk = (chunksize * 8) // bits_per_sample
+    nchunks = samples // samples_per_chunk
+
+    basetypes = {
+        # nbits: (basetype, offset),
+        2: (np.dtype('>u1'), 0),
+        3: (np.dtype('>u4'), 1),
+        4: (np.dtype('>u1'), 0),
+        5: (np.dtype('>u8'), 3),
+        6: (np.dtype('>u4'), 1),
+        7: (np.dtype('>u8'), 1),
+    }
+    basetype, offset = basetypes[bits_per_sample]
+    itemsize = basetype.itemsize
+
+    # byte array
+    npdata = np.frombuffer(data, np.uint8)
+
+    # pad
+    odata = np.zeros([nchunks, itemsize], dtype=np.uint8)
+    odata[:, offset:] = npdata.reshape([nchunks, chunksize])
+
+    # to int
+    odata = odata.view(dtype=basetype)
+
+    # replicate
+    odata = odata.repeat(samples_per_chunk, 1)
+
+    # shift
+    shifts = bits_per_sample * np.arange(samples_per_chunk)[::-1]
+    shifts = np.asarray(shifts, dtype=np.uint32)  # workaround for numpy issue
+    odata = np.right_shift(odata, shifts)
+
+    # mask
+    mask = np.packbits([1] * bits_per_sample, bitorder='little')
+    odata = np.bitwise_and(odata, mask)
+
+    dtype = np.int8 if signed else np.uint8
+    return odata.ravel().astype(dtype)